# specs/001-taboot-rag-platform/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Taboot API
  description: Doc-to-Graph RAG platform API for ingestion, extraction, and retrieval
  version: 1.0.0
  contact:
    name: Taboot Project
    url: https://github.com/yourusername/taboot
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://taboot-app:8000
    description: Docker container (internal)
...
security:
  - bearerAuth: []
paths:
...
  /ingest:
    post:
      tags: [ingestion]
      summary: Start an ingestion job
      description: Ingest documents from specified source with optional limit
      operationId: start_ingestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionRequest'
      responses:
        '202':
          description: Ingestion job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ingest/{job_id}:
    get:
      tags: [ingestion]
      summary: Get ingestion job status
      description: Retrieve current status and progress of an ingestion job
      operationId: get_ingestion_status
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobStatus'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token authentication for API access
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid source type"
        details:
          type: object
          additionalProperties: true
    IngestionRequest:
      type: object
      required:
        - source_type
        - source_target
      properties:
        source_type:
          type: string
          enum: [web, github, reddit, youtube, gmail, elasticsearch, docker_compose, swag, tailscale, unifi, ai_session]
          example: web
        source_target:
          type: string
          description: URL, repo name, file path, or other source identifier
          example: "https://docs.example.com"
        limit:
          type: integer
          description: Maximum pages/items to ingest (optional, defaults to unlimited)
          minimum: 1
          example: 20
    IngestionJobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        state:
          type: string
          enum: [pending, running, completed, failed]
          example: pending
        source_type:
          type: string
          example: web
        source_target:
          type: string
          example: "https://docs.example.com"
        created_at:
          type: string
          format: date-time
          example: "2025-10-21T12:34:56Z"
    IngestionJobStatus:
      allOf:
        - $ref: '#/components/schemas/IngestionJobResponse'
        - type: object
          properties:
            started_at:
              type: string
              format: date-time
              nullable: true
              example: "2025-10-21T12:35:00Z"
            completed_at:
              type: string
              format: date-time
              nullable: true
              example: "2025-10-21T12:36:42Z"
            pages_processed:
              type: integer
              example: 18
            chunks_created:
              type: integer
              example: 342
            errors:
              type: array
              items:
                type: object
                properties:
                  error:
                    type: string
                  url:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
...